requirements.txt
fastapi[all]==0.115.0
uvicorn[standard]==0.30.0
SQLAlchemy==2.0.30
alembic==1.13.2
psycopg2-binary==2.9.9
python-multipart==0.0.9
Jinja2==3.1.4
Pillow==10.3.0
qrcode==7.4.2
pydantic==2.6.4
pydantic-settings==2.2.1
httpx==0.27.0
orjson==3.10.0
xgboost==2.0.3
scikit-learn==1.4.2
pandas==2.2.2
numpy==1.26.4
streamlit
plotly
cachetools
requests
Babel
